% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_utils.R
\name{agg_to_monthly}
\alias{agg_to_monthly}
\title{Aggregate a data frame to monthly frequency}
\usage{
agg_to_monthly(dt, date_name = NULL)
}
\arguments{
\item{dt}{data in data table or data frame format including a date column names "date" or "time"}

\item{date_name}{if the date name is not "date" or "time", specify it here}
}
\description{
Aggregate a data frame to monthly frequency
}
