% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_SA.R
\name{auto_SA}
\alias{auto_SA}
\title{Automatically seasonally adjust high frequency or irregular frequency data}
\usage{
auto_SA(data, dates = NULL, take_logs = "auto", detrend = TRUE)
}
\arguments{
\item{data}{data in matrix format with time in rows}

\item{dates}{dates corresponding to data points if data is not ts_boxable()}
}
\description{
Automatically seasonally adjust high frequency or irregular frequency data
}
