\docType{package}
\name{seasonality}
\alias{seasonality}
\alias{seasonality-package}
\title{seasonality package}
\description{
  This package provides a set of functions to split
  year-round space-use measurements into biological
  seasons, completed with additional functions to explore
  and simplify these seasons. Reference: Basille M., Fortin
  D., Dussault C., Ouellet J.-P., Courtois R. Ecologically
  based definition of seasons clarifies predator-prey
  interactions. Ecography, early view. DOI:
  10.1111/j.1600-0587.2011.07367.x
}
\author{
  Mathieu Basille \email{basille@ase-research.org}
}

