\docType{data}
\name{caribou}
\alias{caribou}
\title{Data set used in Basille et al. (2012).}
\description{
  Data set used in Basille et al. (2012).
}
\author{
  Mathieu Basille \email{basille@ase-research.org}
}
\references{
  Basille M., Fortin D., Dussault C., Ouellet J.-P.,
  Courtois R. Ecologically based definition of seasons
  clarifies predator-prey interactions. Ecography, early
  view. DOI: 10.1111/j.1600-0587.2011.07367.x
}
\keyword{data}
\keyword{datasets}

